requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.2
